refactor(client): Add svelte types to tsconfig, typing `.svelte` files
